修改输入捕获get_zone_time_us()返回值为float类型

+++ ebox/Peripherals/in_capture.cpp
@@ -245,9 +245,9 @@ uint32_t IN_CAPTURE::get_capture()
     last_value = now;
     return capture;
 }
-uint32_t IN_CAPTURE::get_zone_time_us()
+float IN_CAPTURE::get_zone_time_us()
 {
-    return get_capture()/(72/this->prescaler);
+    return get_capture()/(72.0/this->prescaler);
 }
 
 void IN_CAPTURE::set_count(uint16_t count)
